Refactor to comply with code cleanup (#456)
